// src/pages/client/data.d.ts
/*
 * +----------------------------------------------------------------------------------
 * | https://www.tigervs.com
 * +----------------------------------------------------------------------------------
 * | Email: Kevin@tigervs.com
 * +----------------------------------------------------------------------------------
 * | Copyright (c) Shenzhen Tiger Technology Co., Ltd. 2018~2022. All rights reserved.
 * +----------------------------------------------------------------------------------
 */

/** @format */

export type tableDataItem = {
    page?: string;
    email?: string;
    source?: number;
    mobile?: number;
    company?: string;
    message?: string;
    address?: string;
    id?: number | null;
    username?: string;
    ipaddress?: number;
    create_time?: string;
};